Refactored views (#438)
* basic screen

* basics

* adapted separator

* refactored

* refactored

* refactored

* refactored

* fixed test
