+++ backend/src/main/resources/db-migration-testdata/testdata.sql
@@ -12,6 +12,8 @@ INSERT INTO users_authorities (id, created_at, updated_at, user_id, name)
 VALUES (1001, NOW(), NOW(), 100, 'CUSTOMER');
 INSERT INTO users_authorities (id, created_at, updated_at, user_id, name)
 VALUES (1002, NOW(), NOW(), 100, 'SCANNER');
+INSERT INTO users_authorities (id, created_at, updated_at, user_id, name)
+VALUES (1003, NOW(), NOW(), 100, 'CHECKIN');
 
 -- user e2etest2 for cypress test (password change)
 -- pwd: e2etest
@@ -37,6 +39,8 @@ INSERT INTO users_authorities (id, created_at, updated_at, user_id, name)
 VALUES (2001, NOW(), NOW(), 200, 'CUSTOMER');
 INSERT INTO users_authorities (id, created_at, updated_at, user_id, name)
 VALUES (2002, NOW(), NOW(), 200, 'SCANNER');
+INSERT INTO users_authorities (id, created_at, updated_at, user_id, name)
+VALUES (2003, NOW(), NOW(), 200, 'CHECKIN');
 
 -- TODO remove | fake user with weak password to gain some intrusion insights
 -- user: admin
@@ -49,6 +53,8 @@ INSERT INTO users_authorities (id, created_at, updated_at, user_id, name)
 VALUES (3001, NOW(), NOW(), 300, 'CUSTOMER');
 INSERT INTO users_authorities (id, created_at, updated_at, user_id, name)
 VALUES (3002, NOW(), NOW(), 300, 'SCANNER');
+INSERT INTO users_authorities (id, created_at, updated_at, user_id, name)
+VALUES (3003, NOW(), NOW(), 300, 'CHECKIN');
 
 -- user: scanner1
 -- pwd: 12345
