+++ frontend/src/main/webapp/src/app/modules/checkin/checkin-routing.module.ts
@@ -1,11 +1,16 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {ScannerComponent} from './views/scanner/scanner.component';
+import {ScannerComponent} from './scanner/scanner.component';
+import {CheckinComponent} from './checkin/checkin.component';
 
 const routes: Routes = [
   {
     path: 'scanner',
     component: ScannerComponent
+  },
+  {
+    path: 'annahme',
+    component: CheckinComponent
   }
 ];
