+++ frontend/src/main/webapp/src/app/modules/checkin/checkin.module.ts
@@ -1,10 +1,11 @@
 import {NgModule} from '@angular/core';
 import {CheckinRoutingModule} from './checkin-routing.module';
-import {ScannerComponent} from './views/scanner/scanner.component';
+import {ScannerComponent} from './scanner/scanner.component';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {CommonModule} from '@angular/common';
-import {QRCodeReaderService} from './views/scanner/camera/qrcode-reader.service';
+import {QRCodeReaderService} from './qrcode-reader/qrcode-reader.service';
 import {WebsocketService} from '../../common/websocket/websocket.service';
+import {CheckinComponent} from './checkin/checkin.component';
 
 @NgModule({
   imports: [
@@ -14,7 +15,8 @@ import {WebsocketService} from '../../common/websocket/websocket.service';
     FormsModule
   ],
   declarations: [
-    ScannerComponent
+    ScannerComponent,
+    CheckinComponent
   ],
   providers: [
     QRCodeReaderService,
