+++ frontend/src/main/webapp/src/app/modules/checkin/checkin/checkin.component.html
@@ -1,0 +1,9 @@
+<div class="row">
+  <div class="col-md ">
+    <div class="card">
+      <div class="card-body">
+        <h4>Kunden-Annahme</h4>
+      </div>
+    </div>
+  </div>
+</div>
