+++ frontend/src/main/webapp/src/app/modules/checkin/checkin/checkin.component.spec.ts
@@ -1,0 +1,12 @@
+import {TestBed} from '@angular/core/testing';
+import {CheckinComponent} from './checkin.component';
+
+describe('CheckinComponent', () => {
+
+  it('component can be created', () => {
+    const fixture = TestBed.createComponent(CheckinComponent);
+    const component = fixture.componentInstance;
+    expect(component).toBeTruthy();
+  });
+
+});
