+++ frontend/src/main/webapp/src/app/modules/checkin/checkin/checkin.component.ts
@@ -1,0 +1,8 @@
+import {Component} from '@angular/core';
+
+@Component({
+  selector: 'tafel-checkin',
+  templateUrl: 'checkin.component.html'
+})
+export class CheckinComponent {
+}
