+++ frontend/src/main/webapp/src/app/modules/checkin/qrcode-reader/qrcode-reader.service.ts
