+++ frontend/src/main/webapp/src/app/modules/checkin/scanner/scanner.component.html
