+++ frontend/src/main/webapp/src/app/modules/customer/customer-routing.module.ts
@@ -1,9 +1,9 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {CustomerDetailComponent} from './views/customer-detail.component';
+import {CustomerDetailComponent} from './views/customer-detail/customer-detail.component';
 
-import {CustomerEditComponent} from './views/customer-edit.component';
-import {CustomerSearchComponent} from './views/customer-search.component';
+import {CustomerEditComponent} from './views/customer-edit/customer-edit.component';
+import {CustomerSearchComponent} from './views/customer-search/customer-search.component';
 
 const routes: Routes = [
   {
