+++ frontend/src/main/webapp/src/app/modules/customer/customer.module.ts
@@ -1,15 +1,15 @@
 import {NgModule} from '@angular/core';
 import {ReactiveFormsModule} from '@angular/forms';
 
-import {CustomerEditComponent} from './views/customer-edit.component';
+import {CustomerEditComponent} from './views/customer-edit/customer-edit.component';
 import {CustomerRoutingModule} from './customer-routing.module';
-import {CustomerFormComponent} from './components/customer-form.component';
+import {CustomerFormComponent} from './views/customer-form/customer-form.component';
 import {CommonModule} from '@angular/common';
 import {ModalModule} from 'ngx-bootstrap/modal';
-import {CustomerDetailComponent} from './views/customer-detail.component';
-import {CustomerSearchComponent} from './views/customer-search.component';
+import {CustomerDetailComponent} from './views/customer-detail/customer-detail.component';
+import {CustomerSearchComponent} from './views/customer-search/customer-search.component';
 import {BsDropdownModule} from 'ngx-bootstrap/dropdown';
-import {CustomerApiService} from './api/customer-api.service';
+import {CustomerApiService} from '../../api/customer-api.service';
 
 @NgModule({
   imports: [
