+++ frontend/src/main/webapp/src/app/modules/customer/views/customer-detail/customer-detail.component.html
