+++ frontend/src/main/webapp/src/app/modules/customer/views/customer-detail/customer-detail.component.spec.ts
@@ -5,8 +5,8 @@ import {ActivatedRoute, Router} from '@angular/router';
 import {RouterTestingModule} from '@angular/router/testing';
 import * as moment from 'moment';
 import {of} from 'rxjs';
-import {FileHelperService} from '../../../common/util/file-helper.service';
-import {CustomerApiService, CustomerData} from '../api/customer-api.service';
+import {FileHelperService} from '../../../../common/util/file-helper.service';
+import {CustomerApiService, CustomerData} from '../../../../api/customer-api.service';
 import {CustomerDetailComponent} from './customer-detail.component';
 import {CommonModule} from '@angular/common';
 import {DEFAULT_CURRENCY_CODE, LOCALE_ID} from '@angular/core';
