+++ frontend/src/main/webapp/src/app/modules/customer/views/customer-search/customer-search.component.html
@@ -24,9 +24,9 @@
                         [disabled]="!customerId.value" (click)="searchForCustomerId()">Anzeigen</button>
                 </div>
             </div>
-            <div class="row mt-4 mb-3">
-                <div class="col-5">
-                    <h5>Detailsuche</h5>
+            <div class="row">
+                <div class="col-6">
+                    <hr/>
                 </div>
             </div>
             <div class="row">
