+++ frontend/src/main/webapp/src/app/modules/dashboard/navigation-menuItems.ts
@@ -17,6 +17,12 @@ export const navigationMenuItems: IPermissionNavData[] = [
     title: true
   },
   {
+    name: 'Annahme',
+    url: '/anmeldung/annahme',
+    icon: 'fa fa-check',
+    permissions: ['CHECKIN']
+  },
+  {
     name: 'Scanner',
     url: '/anmeldung/scanner',
     icon: 'fa fa-barcode',
