+++ frontend/src/main/webapp/src/app/modules/user/views/passwordchange-form/passwordchange-form.component.ts
@@ -1,6 +1,6 @@
 import {Component} from '@angular/core';
 import {FormControl, FormGroup, ValidatorFn, Validators} from '@angular/forms';
-import {ChangePasswordRequest, ChangePasswordResponse, UserApiService} from '../../api/user-api.service';
+import {ChangePasswordRequest, ChangePasswordResponse, UserApiService} from '../../../../api/user-api.service';
 import {catchError, map} from 'rxjs/operators';
 import {Observable, throwError} from 'rxjs';
 import {HttpErrorResponse} from '@angular/common/http';
